Create arrow_game_sensehat.py
This is a game based on sensehat. On the sense hat, a green arrow that has a random direction will be displayed. Player need to change the direction of the arrow to up in 10 seconds, if they do so, game continue and the arrow of another direction is generated, otherwise, arrow becomes red and game over.